Range Sensor Method Fix v2
The Range Sensor Method now works almost perfectly.

There is some rubber band due to the motors being to powerful or too weak, but it ends up in the right spot. We need to add a PID for the range between the inAway and sensor value so it doesn't do this.