Fixed A lot of
Vuforia was tampered with so spent time fixing that. The error was a simple issue of not placing the variable update in the right spot although I am unsure if this works with column (I'd assume so).

TurnCorr was optimized in runtime and code which seemed to be redundant.

There are still changes being made to everything else.