TurnCorr Fix - Turns Closer
The robot used to have to make a 180 degree circle if it overshot the first turn and turned over the marker where it would become negative (i.e from 360 to 0) in a circle. Changed the way it calculates how to move so now it moves to the distance closest to it.

Still need to change auto files.